+++ crates/miden-protocol/src/transaction/mod.rs
@@ -29,6 +29,7 @@ pub use logs::{
     LogTopic,
     TransactionLog,
     TransactionLogData,
+    TransactionLogDataCollection,
     TransactionLogDataError,
     TransactionLogs,
 };